app/requirements.txt
accelerate==0.26.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.1.0
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.11.1
bitsandbytes==0.41.2.post2
bleach==5.0.1
blinker==1.7.0
blis==0.7.11
build==1.1.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.4
coremltools==6.3.0
cryptography==42.0.5
cymem==2.0.8
Cython==0.29.37
dataclasses-json==0.6.4
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diskcache==5.6.3
distlib==0.3.6
distro==1.9.0
entrypoints==0.4
executing==1.2.0
faiss-cpu==1.8.0
fastapi==0.110.0
fastjsonschema==2.16.2
filelock==3.12.0
flatbuffers==24.3.7
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.29.0
googleapis-common-protos==1.63.0
gptcache==0.1.43
greenlet==3.0.3
grpcio==1.62.1
guidance==0.1.10
h11==0.14.0
hdbscan==0.8.33
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.11.0
importlib_resources==6.4.0
IProgress==0.4
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab_widgets==3.0.10
kubernetes==29.0.0
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-text-splitters==0.0.1
langcodes==3.3.0
langsmith==0.1.31
llama_cpp_python==0.2.26
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.28.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.59.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.14.3
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
ordered-set==4.1.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
peft==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pip-system-certs==4.0
platformdirs==3.5.1
plotly==5.20.0
posthog==3.5.0
preshed==3.0.9
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.3
ptyprocess==0.7.0
PuLP==2.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pydot==2.0.0
pyformlang==1.0.9
Pygments==2.13.0
PyJWT==2.8.0
pynndescent==0.5.11
pyparsing==3.0.9
PyPika==0.48.9
pyproject_hooks==1.0.0
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.6
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.2.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
Send2Trash==1.8.0
sentence-transformers==2.5.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
stack-data==0.6.0
starlette==0.36.3
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
terminado==0.17.0
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.5.0
transformers==4.37.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
umap-learn==0.5.5
unidecode==1.3.8
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.23.0
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.5
weasel==0.3.4
webencodings==0.5.1
websocket-client==1.4.1
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
